requirements.txt
anthropic
python-dotenv
streamlit
newspaper3k
lxml_html_clean
